### 2024-06-27 #6935 @gavinsimpson
Suggestions for reviewers from list:

* vankesteren
* njtierney
* epiben

Suggestions for reviewers

* Andreas Bender (adibender) - maintainer of *pammtools* an R package for fitting and visualizing event history models fitted as GAMs
* Nick Golding (goldingn) would be suitable too

Depending on the statute of limitations on CoIs from having coauthored a paper 5-ish years ago, obvious reviewers would be

* Dave Miller (dill)
* Noam Ross (noamross)

But they might decline anyway. Paper we published: https://doi.org/10.7717/peerj.6876
