### 2024-06-27 #6935 @gavinsimpson
> 🟡 License found: Other ([Check here](https://opensource.org/licenses) for OSI approval)

The package is licensed MIT: https://github.com/gavinsimpson/gratia/blob/main/LICENSE.md. That this isn't clearer in the repo is because of the requirements for conveying license info in R packages that are on CRAN.
