### 2024-06-27 #6935 @lrnv
@vankesteren, @njtierney, @epiben, @adibender, @goldingn, @dill, @noamross, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
