### 2024-06-27 68f5ab2
move some code from fitted_values
