### 2024-06-27 e376b28
fixes #289; adds dispersion() and n_eta() utilities
