+++ man/dispersion.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utililties.R
+\name{dispersion}
+\alias{dispersion}
+\alias{dispersion.gam}
+\alias{dispersion.glm}
+\title{Dispersion parameter for fitted model}
+\usage{
+dispersion(model, ...)
+
+\method{dispersion}{gam}(model, ...)
+
+\method{dispersion}{glm}(model, ...)
+}
+\arguments{
+\item{model}{a fitted model.}
+
+\item{...}{arguments passed to other methods.}
+}
+\description{
+\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
+}
