+++ man/post_link_funs.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/fitted_values.R
+% Please edit documentation in R/distributional-models.R
 \name{post_link_funs}
 \alias{post_link_funs}
 \title{A list of transformation functions named for LSS parameters in a GAMLSS}
