+++ man/lss_parameters.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/fitted_values.R
+% Please edit documentation in R/distributional-models.R
 \name{lss_parameters}
 \alias{lss_parameters}
 \title{General names of LSS parameters for each GAM family}
