+++ NAMESPACE
@@ -58,6 +58,8 @@ S3method(difference_smooths,bam)
 S3method(difference_smooths,gam)
 S3method(difference_smooths,gamm)
 S3method(difference_smooths,list)
+S3method(dispersion,gam)
+S3method(dispersion,glm)
 S3method(draw,basis)
 S3method(draw,compare_smooths)
 S3method(draw,derivatives)
@@ -145,12 +147,15 @@ S3method(link,list)
 S3method(lp_matrix,gam)
 S3method(mh_draws,gam)
 S3method(model_concurvity,gam)
+S3method(model_constant,gam)
+S3method(model_constant,gamlss)
 S3method(model_vars,bam)
 S3method(model_vars,default)
 S3method(model_vars,gam)
 S3method(model_vars,gamm)
 S3method(model_vars,gamm4)
 S3method(model_vars,list)
+S3method(n_eta,gam)
 S3method(n_smooths,bam)
 S3method(n_smooths,default)
 S3method(n_smooths,gam)
@@ -318,6 +323,7 @@ export(data_slice)
 export(derivative_samples)
 export(derivatives)
 export(difference_smooths)
+export(dispersion)
 export(draw)
 export(edf)
 export(eval_smooth)
@@ -361,6 +367,7 @@ export(model_concurvity)
 export(model_constant)
 export(model_edf)
 export(model_vars)
+export(n_eta)
 export(n_smooths)
 export(nb_theta)
 export(null_deviance)
