### 2024-06-27 4a5f574
add a new model; update test for model_constant
