+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   data-slices: data-slices.html
   gratia: gratia.html
   posterior-simulation: posterior-simulation.html
-last_built: 2024-06-26T17:46Z
+last_built: 2024-06-27T18:01Z
 urls:
   reference: https://gavinsimpson.github.io/gratia/reference
   article: https://gavinsimpson.github.io/gratia/articles
