### 2024-06-28 6a8a1af
probably best to use all the terms even though the distributional etas don't affect the expectation - they would if this was posterior_samples()
