### 2024-06-28 aec0304
in hindsight, simplifying the code is best; added a note in the text now to say I'm also excluding terms in the other linear predictors.
