### 2024-06-28 4ce504b
implement one part of #288; turn the plot for soap film smooths into an isotropic one
