+++ NAMESPACE
@@ -454,6 +454,7 @@ importFrom(ggokabeito,scale_colour_okabe_ito)
 importFrom(ggokabeito,scale_fill_okabe_ito)
 importFrom(ggplot2,aes)
 importFrom(ggplot2,coord_equal)
+importFrom(ggplot2,coord_fixed)
 importFrom(ggplot2,coord_sf)
 importFrom(ggplot2,expand_limits)
 importFrom(ggplot2,facet_grid)
