+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   data-slices: data-slices.html
   gratia: gratia.html
   posterior-simulation: posterior-simulation.html
-last_built: 2024-06-27T18:01Z
+last_built: 2024-06-28T15:41Z
 urls:
   reference: https://gavinsimpson.github.io/gratia/reference
   article: https://gavinsimpson.github.io/gratia/articles
