### 2024-06-29 #6935 @goldingn
Sorry @lrnv, I haven't done a JOSS review in ages so would love to, but I will be slammed for at least the next couple of months :(
