### 2024-07-04 #6962 @editorialbot
[REVIEW]: gratia: An R package for exploring generalized additive models
**Submitting author:** @gavinsimpson (<a href="http://orcid.org/0000-0002-9084-8413">Gavin L. Simpson</a>)
**Repository:** https://github.com/gavinsimpson/gratia
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.10.0
