### 2024-07-04 #6962 @lrnv
@dill, @vankesteren, @gavinsimpson this is the review thread for the paper. All of our communications will happen here (and on issues/prs on the submitted repository) from now on.

@dill, @vankesteren, as reviewers, the first step for each of you is to create a checklist for your review by entering

as the top of a new comment in this thread. These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention openjournals/joss-reviews#6962 so that a link is created to this thread (and I can keep an eye on what is happening !). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
