### 2024-07-15 #6962 @dill
In terms of conflict of interest: I've worked with @gavinsimpson several times, having written a paper together and taught two courses. I've also made a contribution to `gratia`. I don't feel like this will have a material impact on my ability to review the paper -- I'm invested in the package being the best it can be, this requires constructive criticism.
