### 2024-07-15 #6962 @dill
Comments on the paper are given in https://github.com/gavinsimpson/gratia/issues/296.
