### 2024-07-15 gavinsimpson/gratia#296 @gavinsimpson
> The author name is shown as "¿citation_author?" in the pdf on the gratia github. Not sure if this is normal.

This seems to be OK in version generated by the review system itself (I think I have something wrong in the YAML meta data for the the `citation_author` element which is causing the problem there.) So I'll leave this just now (beyond tweaking the `citation_author` in the YAML) and figure it out if the issue presents later.
