### 2024-07-16 gavinsimpson/gratia#296 @gavinsimpson
> Going back to the boring situation in which we "read the documentation", in ?smooth.construct we have:

>> null.space.dim: The dimension of the penalty null space (before centering).
So I think this is before identifiability constraints are applied (since they need to be applied at the model, not term level. So I think the nullspace just consists of the intercept in that case.

But for that to be true, wouldn't the null space dimension for a TPRS (the default) be = 2 because it has both constant and linear basis functions, yet as you show it is 1.

If we do

```r
library("mgcv")
library("gratia")

df <- data_sim("gwf2", seed = 2)
m <- gam(y ~ s(x, bs = "cr"), data = df)

# extract penalty
S <- m$smooth[[1]]$S[[1]]

# eigen decompoisition
eigen(S)
```

producing

```
> eigen(S)
eigen() decomposition
$values
[1]  2.201890e+00  1.420716e+00  8.582862e-01  4.449231e-01  1.969655e-01
[6]  7.046850e-02  1.825633e-02  2.373200e-03 -6.021430e-17
```

where 1 eigenvalue is numerically 0 and the matrix is not invertible by normal means.

```r
$> solve(S)
Error in solve.default(S) :
  system is computationally singular: reciprocal condition number = 4.66992e-19
```

And this is what I would expect, because a linear function is still in the span of basis, even if there is no actual linear function.

Now consider the same model but with `select = TRUE`, where we include an additional penalty on the null space:

```r
m <- gam(y ~ s(x, bs = "cr"), data = df, select = TRUE)
```

Now the null space of this smooth is 0, as it should be because of the extra penalty:

```
> m$smooth[[1]]$null.space.dim
[1] 0
```

Some more discussion is needed here...
