### 2024-07-16 gavinsimpson/gratia#296 @gavinsimpson
Some further reading has helped, if we're just talking about the Null space of a penalty. The key point I think is that the null space is size of the span of functions representable by the basis for which the penalty has no effect. The null space doesn't refer to specific basis functions but to the functions representable by that basis.

Section 5.4.3 of Simon's GAM book has:

![BB0B45E9-91FB-4F35-B194-73D0F06C1446](https://github.com/user-attachments/assets/703c3fa7-f93b-43d9-be03-5e681164933f)

which I think is clear. This explains the findings in my comment above. I have assumed that the rank deficiency of $S$ is what leads to the improper prior when we view penalized smoothing from a Bayesian perspective. But I should read a little more about that...
