+++ NAMESPACE
@@ -227,7 +227,7 @@ S3method(smooth_terms,gam)
 S3method(smooth_terms,gamm)
 S3method(smooth_terms,mgcv.smooth)
 S3method(smooth_type,Bspline.smooth)
-S3method(smooth_type,cp.smooth)
+S3method(smooth_type,cpspline.smooth)
 S3method(smooth_type,cr.smooth)
 S3method(smooth_type,cs.smooth)
 S3method(smooth_type,cv.smooth)
