+++ tests/testthat/_snaps/soap-films.md
@@ -4,17 +4,17 @@
       print(sm_so)
     Output
       # A tibble: 10,160 x 9
-         .smooth .type     .by   .estimate   .se .bndry .loop      v      w
-         <chr>   <chr>     <chr>     <dbl> <dbl> <lgl>  <int>  <dbl>  <dbl>
-       1 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.898
-       2 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.880
-       3 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.862
-       4 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.844
-       5 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.826
-       6 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.808
-       7 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.789
-       8 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.771
-       9 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.753
-      10 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.886 -0.735
+         .smooth .type     .by   .estimate   .se .bndry .loop      v     w
+         <chr>   <chr>     <chr>     <dbl> <dbl> <lgl>  <int>  <dbl> <dbl>
+       1 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.9    -0.9
+       2 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.857  -0.9
+       3 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.813  -0.9
+       4 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.770  -0.9
+       5 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.726  -0.9
+       6 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.683  -0.9
+       7 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.639  -0.9
+       8 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.596  -0.9
+       9 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.553  -0.9
+      10 s(v,w)  Soap film <NA>         NA    NA FALSE     NA -0.509  -0.9
       # i 10,150 more rows
 
