+++ tests/testthat/_snaps/central/response-derivatives.md
@@ -26,15 +26,15 @@
       # A tibble: 50 x 9
           .row .focal .derivative .lower_ci .upper_ci    x0    x1      x2    x3
          <int> <chr>        <dbl>     <dbl>     <dbl> <dbl> <dbl>   <dbl> <dbl>
-       1     1 x2            217.     116.       316. 0.480 0.478 0.00127 0.537
-       2     2 x2            276.     163.       407. 0.480 0.478 0.0216  0.537
-       3     3 x2            374.     264.       525. 0.480 0.478 0.0420  0.537
-       4     4 x2            496.     399.       633. 0.480 0.478 0.0624  0.537
-       5     5 x2            614.     487.       750. 0.480 0.478 0.0828  0.537
-       6     6 x2            683.     464.       898. 0.480 0.478 0.103   0.537
-       7     7 x2            596.     281.       897. 0.480 0.478 0.123   0.537
-       8     8 x2            352.     -37.1      734. 0.480 0.478 0.144   0.537
-       9     9 x2           -142.    -550.       252. 0.480 0.478 0.164   0.537
-      10    10 x2           -842.   -1258.      -461. 0.480 0.478 0.185   0.537
+       1     1 x2            209.     105.       309. 0.480 0.478 0.00127 0.537
+       2     2 x2            291.     186.       416. 0.480 0.478 0.0216  0.537
+       3     3 x2            379.     270.       528. 0.480 0.478 0.0420  0.537
+       4     4 x2            505.     406.       642. 0.480 0.478 0.0624  0.537
+       5     5 x2            616.     488.       753. 0.480 0.478 0.0828  0.537
+       6     6 x2            690.     479.       900. 0.480 0.478 0.103   0.537
+       7     7 x2            625.     274.       963. 0.480 0.478 0.123   0.537
+       8     8 x2            345.     -61.0      744. 0.480 0.478 0.144   0.537
+       9     9 x2           -166.    -576.       227. 0.480 0.478 0.164   0.537
+      10    10 x2           -856.   -1295.      -464. 0.480 0.478 0.185   0.537
       # i 40 more rows
 
