### 2024-07-17 a90abe2
document fix for #297 and bump version
