+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: gratia
-Version: 0.9.2.9002
-Date: 2024-06-26
+Version: 0.9.2.9003
+Date: 2024-07-17
 Title: Graceful 'ggplot'-Based Graphics and Other Functions for GAMs Fitted Using 'mgcv'
 Authors@R: c(person(given = "Gavin L.", family = "Simpson",
                     email = "ucfagls@gmail.com",
