+++ pkgdown.yml
@@ -1,13 +1,12 @@
 pandoc: 3.1.11
-pkgdown: 2.0.9
+pkgdown: 2.1.0
 pkgdown_sha: ~
 articles:
-  custom-plotting: custom-plotting.html
+  articles/custom-plotting: custom-plotting.html
   data-slices: data-slices.html
   gratia: gratia.html
-  posterior-simulation: posterior-simulation.html
-last_built: 2024-06-28T15:41Z
+  articles/posterior-simulation: posterior-simulation.html
+last_built: 2024-07-17T09:01Z
 urls:
   reference: https://gavinsimpson.github.io/gratia/reference
   article: https://gavinsimpson.github.io/gratia/articles
-
