+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   data-slices: data-slices.html
   gratia: gratia.html
   articles/posterior-simulation: posterior-simulation.html
-last_built: 2024-07-17T09:01Z
+last_built: 2024-07-17T09:06Z
 urls:
   reference: https://gavinsimpson.github.io/gratia/reference
   article: https://gavinsimpson.github.io/gratia/articles
