### 2024-07-17 c1febab
try to get better data for soap films
