### 2024-07-17 abd2e67
add ability to diagonalize the penalty to basis and penalty
