### 2024-07-17 23eb93f
document new diagonalize argument and bump version
