+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: gratia
-Version: 0.9.2.9003
+Version: 0.9.2.9004
 Date: 2024-07-17
 Title: Graceful 'ggplot'-Based Graphics and Other Functions for GAMs Fitted Using 'mgcv'
 Authors@R: c(person(given = "Gavin L.", family = "Simpson",
