### 2024-07-17 9aae5a7
add tests for new diagonalize argument
