### 2024-07-23 4fadaf8
fix typo; closes #298
