### 2024-07-27 f027793
adjust test for #301
