### 2024-07-27 451e293
document bug fix for #301
