+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: gratia
-Version: 0.9.2.9004
-Date: 2024-07-17
+Version: 0.9.2.9005
+Date: 2024-07-27
 Title: Graceful 'ggplot'-Based Graphics and Other Functions for GAMs Fitted Using 'mgcv'
 Authors@R: c(person(given = "Gavin L.", family = "Simpson",
                     email = "ucfagls@gmail.com",
