+++ tests/testthat/test-derivatives.R
@@ -478,9 +478,9 @@ test_that("derivatives() works for fs smooths issue 57 and 301", {
 
   # test variant in #301
   logistic_growth_df$linearterm = runif(16)
-  m <- gam(y_obs ~ s(t, unit, k = 5, bs = "fs", m = 2) + linearterm,
+  m_logistic_growth <- gam(y_obs ~ s(t, unit, k = 5, bs = "fs", m = 2) + linearterm,
     data = logistic_growth_df, method = "REML")
-  expect_silent(m <- derivatives(m))
+  expect_silent(fd <- derivatives(m_logistic_growth))
 })
 
 ## tests for by variables & simultaneous intervals #102
