### 2024-07-29 gavinsimpson/gratia#296 @dill
After some extensive chat here, online and offline (by nice coincidence at the same conference), we've got somewhere on this (the improper prior business)... I'll write this collectively for both our findings but if @gavinsimpson needs to correct me anywhere, I trust that will happen.

- if we have a 2nd order derivative penalty, functions of the form intercept + slope are in the null space of the penalty
- that *doesn't* mean that intercept + slope have to be explicitly in the basis definition, just that whenever the spline looks like a slope, it will not be penalized.
  - you can observe this by taking a B-spline basis (which doesn't have a slope term) and fitting a model where you specify a numerically infinite smoothing parameter (`sp=10e10`, e.g.) -- we have a straight line, even though we don't have an explicit slope term
- when we absorb the intercept for each term into the "global" intercept (which we must do for identifiability in the non-fully Bayesian, informative priors case) we remove that intercept from each term, BUT there still remains the slope null space
- improper priors occur when we have non-finite variances (amongst other reasons), these occur when we have approximately zero penalties (since the covariance is the inverse of the penalty, large variance -> small penalty) which occurs when we have a rank deficient penalty -- when there is a null space.
- there is a little subtlety about whether we talk about the prior for an individual smooth term vs. the multivariate normal prior for the whole model (taking a kind of Gaussian process view of the model, each bit is Gaussian, so the whole is one big Gaussian).
- there are some limited cases where we have a proper prior on individual smooths, for example:
  - when using the `ts` and `cs` bases (where there is a small penalty added to the null space terms)
  - when using `select=TRUE` additional penalty matrices for the null space terms
  - when using a lower-order penalty, with certain bases (`tp`, `ps`, `bs`, maybe others), where the slope is not in the null space and the intercept is removed with absorbing constraints.

Does that cover everything @gavinsimpson?
