### 2024-07-29 gavinsimpson/gratia#296 @dill
Then two additional comments because @gavinsimpson mistakenly asked me to look more closely at his equations...

- Generally, I think folks use $\mathcal{L}$ to be the likelihood at define $\ell =  \log_e\mathcal{L}$, so it might be "better" to write your $\ell$ as $\ell_p$ or somesuch?

(this list interrupted by my realising that github now renders $\LaTeX$, wow! That's going to make the next bit easier...)

- on combining the $\boldsymbol{\gamma}$ and $\boldsymbol{\beta}_j$ into $\boldsymbol{\beta}$: I think this is fine, you can say you do that and say that $\boldsymbol{\beta}_j$ refers to the appropriate parameters relating to term $j$ in the model, then I think that's clear enough.
