### 2024-07-29 gavinsimpson/gratia#302 @dill
JOSS review: Community guidelines
Part of the JOSS review [here](https://github.com/openjournals/joss-reviews/issues/6962) says asks "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I see the code of conduct, but I don't see anything about contributions. Perhaps a wee section in the `README.md` like `jagstargets` (for example) does [here](https://github.com/ropensci/jagstargets?tab=readme-ov-file#participation) would be helpful?
