### 2024-07-30 5d4ea35
avoid calls to sapply
