### 2024-08-05 gavinsimpson/gratia#304 @vankesteren
JOSS review: paper comments
congrats on your submission to JOSS! This is an issue coming out of the JOSS review [here](https://github.com/openjournals/joss-reviews/issues/6962). This looks like a great & very useful package. Here is a checklist for points relating to your paper.

- [x] I have a feeling that the statement of need could be improved to instil an increased sense of urgency in the reader. What is the current problem with GAMs which is solved by {gratia}? Who will use it? The last paragraph of that section (starting line 82) is what I actually want to read much earlier, maybe even in the summary section.
- [ ] Simplify package loading in your example usage in the paper. Maybe the following would be easier to parse for newcomers? `for (pkg in pkgs) library(pkg, character.only = TRUE)`
- [x] In the paper, add a code comment before the line `crs <- "+proj=ortho +lat_0=20 +lon_0=-40"` for accessibility purposes. Something like: # Define the coordinate system to use for plotting on a map.
These geo coordinate reference system codes are not clear to newcomers!
- [x] In the paper and "getting started" vignette alike, it would be nice to have a short explanation of what a "partial effect" means. What are we actually seeing in the plots? Is that the smooth evaluated at the average of all other variables, is it evaluated marginalizing out the other variables, or is some other reference chosen? Some engagement with, e.g., the marginaleffects package (Arel-Bundock, Greifer, & Heiss, Forthcoming) might be appropriate here. This relates to the next point
- [x] State of the field: in the JOSS review checklist, it says "Do the authors describe how this software compares to other commonly-used packages?". This is currently not the case, so a section for this needs to be added.
- [x] A bit nitpicky, but consider making the following language more accessible for newcomers
  - [x] L132: "which is easily plotted using" (easy for experts, not for newcomers)
  - [x] l135: "quite a simple matter to calculate" (simple for experts, not for newcomers)
  - [x] L159: "While it would be a simple matter to compute the interval with base R commands" (idem)
- [x] Space before comma on line 150
- [x] Maybe consider adding a very short 1 paragraph conclusion to your article (personal preference, feel free to ignore / check off)

Arel-Bundock V, Greifer N, Heiss A (Forthcoming). “How to Interpret Statistical Models Using marginaleffects in R and Python.” Journal of Statistical Software. https://marginaleffects.com
