### 2024-08-05 gavinsimpson/gratia#305 @vankesteren
JOSS review: code points
Hey @gavinsimpson,

congrats on your submission to JOSS! This is an issue coming out of the JOSS review [here](https://github.com/openjournals/joss-reviews/issues/6962). Here is a checklist for points relating to your code / documentation.

- [ ] The readme could benefit from a better statement of need in the overview section. Short, to the point, basically "gams can be difficult to interpret, gratia helps with this".
- [x] Why `draw()` and why not the ggplot2 built-in generic [`autoplot()`](https://ggplot2.tidyverse.org/reference/autoplot.html)? That is for example used in packages like [`fpp3`](https://otexts.com/fpp3/time-plots.html). I feel like adding an additional generic is maybe overkill? You could have also overloaded `plot.gam()` (although I do appreciate the issues with that so that may be a bad idea). I also realize that your code is already widely used so that may hinder changing something so fundamental about your API.
- [x] Documentation for the argument "method" in the function `appraise()` is incomplete. Please document the other methods! (e.g., simulate, which is included in the paper as an example)
- [x] Fix the FIXME statement in the documentation of the data argument in `draw.gam()`.
- [x] The function `data_slice()` should be better documented, to avoid confusion with another common use of the term "slice" as indexing (e.g., https://stackoverflow.com/questions/509211/how-slicing-in-python-works, https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice). How are the values for the non-focal covariates computed? are they averages? are they local averages? (they seem to be averages)

I haven't yet had the time to go over the other documentation (it's a lot!), but will try to do so ASAP.
