### 2024-08-05 e594bca
Fix the FIXME statement in the documentation of the data argument; #305
