### 2024-08-05 03377ff
Fixes: Documentation for the argument "method" in the function appraise() is incomplete; #305
