### 2024-08-05 bf2ad20
The function data_slice() should be better documented: #305
