### 2024-08-05 21d2dae
The readme could benefit from a better statement of need in the overview section; #305
