### 2024-08-05 gavinsimpson/gratia#305 @gavinsimpson
@vankesteren I have addressed each of the remaining points. If the changes do not go far enough, reopen this and I can take another stab.
