### 2024-08-05 gavinsimpson/gratia#304 @gavinsimpson
Re

> Simplify package loading in your example usage in the paper. Maybe the following would be easier to parse for newcomers?

I don't really agree here; while the suggested change might be easier to understand for newcomers, I don't see why we can't use pretty basic language features. And as most people (certainly the intended users gratia) probably don't understand `for()` loops anyway, I don't see the issue with using `vapply()`.
