### 2024-08-10 90c4d78
allow captions to be turned off with  in ; wish of #307 in discussion #306
