### 2024-08-11 7d5947c
allow user to supply seed for qq_plot() and thence appraise()
