### 2024-08-12 #6962 @dill
Sorry for the late reply here, I concluded my review with comments listed above in the `gratia` repo. Once those are resolved, I'm happy with the package!
