### 2024-08-26 #6962 @lrnv
@njtierney, did you have the time to take a look at the package yet?

@gavinsimpson, how is the review going so far? It looks like the discussions are extensive and interesting from here, but I wanted to ask your opinion!
