### 2024-09-02 62232da
Fix the workflow so that it compacts vignettes as needed
