### 2024-10-19 680b089
need working weights for weighted working residuals; see #273
