### 2024-10-19 6fcf0e8
document fix for #273 - use working weights in partial residuals
