### 2024-10-31 d46dc77
gammals should take theta and phi as names for the scale parameter as Simon uses both of those
