### 2024-10-31 8017643
document fix of bug for gammals scale parameter synonym issue
