### 2024-11-21 #6962 @gavinsimpson
@lrnv consider me gently-bumped :)

(Teaching ends next week so I'll have time to pick this up and finish off my response / fixes then.)
