### 2024-11-21 2a22ba3
add a contributing guide as per @vankesteren's checklist at openjournals/joss-reviews#6962
