### 2024-11-21 c078679
address @vankesteren review comment re urgency; move a statement to the summary #304
