+++ paper.Rmd
@@ -27,7 +27,7 @@ journal: JOSS
 
 Generalized additive models [GAMs, @Hastie1990-bx; @Wood2017-qi] are an extension of the generalized linear model that allows the effects of covariates to be modelled as smooth functions. GAMs are increasingly used in many areas of science [e.g. @Pedersen2019-ff; @Simpson2018-wc] because the smooth functions allow nonlinear relationships between covariates and the response to be learned from the data through the use of penalized splines. Within the R [@rcore2024] ecosystem, Simon Wood's  *mgcv* package [@Wood2017-qi] is widely used to fit GAMs and is a *Recommended* package that ships with R as part of the default install. A growing number of other R packages build upon *mgcv*, for example as an engine to fit specialised models not handled by *mgcv* itself [e.g. *GJMR*, @Marra2023-gjrm], or to make use of the wide range of splines available in *mgcv* [e.g. *brms*, @Burkner2017-ms].
 
-The *gratia* package builds upon *mgcv* by providing functions that make working with GAMs easier. *gratia* takes a *tidy* approach [@Wickham2014-ev] providing *ggplot2* [@Wickham2016-dg] replacements for *mgcv*'s base graphics-based plots, functions for model diagnostics and exploration of fitted models, and a family of functions for drawing samples from the posterior distribution of a fitted GAM. Additional functionality is provided to facilitate the teaching and understanding of GAMs.
+The *gratia* package builds upon *mgcv* by providing functions that make working with GAMs easier. *gratia* takes a *tidy* approach [@Wickham2014-ev] providing *ggplot2* [@Wickham2016-dg] replacements for *mgcv*'s base graphics-based plots, functions for model diagnostics and exploration of fitted models, and a family of functions for drawing samples from the posterior distribution of a fitted GAM. Additional functionality is provided to facilitate the teaching and understanding of GAMs. The overall aim of *gratia* is to abstract away some of the complexity of working with GAMs fitted using *mgcv* to allow researchers to focus on using and interrogating their model rather than the technical R programming needed to achieve this.
 
 # Generalized additive models
 
