+++ paper.Rmd
@@ -97,13 +97,13 @@ p_bfs + p_S + plot_layout(ncol = 2, nrow = 1) +
 
 # Statement of need
 
-*mgcv* is state-of-the-art software for fitting GAMs and their extensions to data sets on the order of millions of observations [e.g. @Li2020-ch; @Wood2011-kn; @Wood2016-fx]. *mgcv* provides functions for plotting estimated smooth functions, as well as for producing model diagnostic plots. These functions produce plots using base graphics, the original plotting system for R. Additionally, *mgcv* returns fitted GAMs as complex list objects (see `?mgcv::gamObject`), the contents of which are not easily used for downstream analysis without careful study of *mgcv* and its help pages, plus a good understanding of GAMs themselves. One of the motivations driving the development of *gratia* was to provide equivalent plotting capabilities using the *ggplot2* package [@Wickham2016-dg]. To facilitate this, *gratia* provides functions for representing the model components as objects using *tidy* principles, which are suitable for plotting with *ggplot2* or manipulation by packages in the *tidyverse* [e.g. @Wickham2023-uj]. This functionality allows for high-level plotting using the `draw()` method, as well as easily customisable plot creation using lower-level functionality.
+*mgcv* is state-of-the-art software for fitting GAMs and their extensions to data sets on the order of millions of observations [e.g. @Li2020-ch; @Wood2011-kn; @Wood2016-fx]. *mgcv* provides functions for plotting estimated smooth functions, as well as for producing model diagnostic plots. These functions produce plots using base graphics, the original plotting system for R. Additionally, *mgcv* returns fitted GAMs as complex list objects (see `?mgcv::gamObject`), the contents of which are not easily used for downstream analysis without careful study of *mgcv* and its help pages, plus a good understanding of GAMs themselves. The overall aim of *gratia* is to abstract away some of the complexity of working with GAMs fitted using *mgcv* to allow researchers to focus on using and interrogating their model rather than the technical R programming needed to achieve this. As a result, *gratia* is also increasingly being used by researchers in many fields, and has, at the time of writing, been cited over 250 times (data from Google Scholar).
+
+One of the motivations driving the development of *gratia* was to provide equivalent plotting capabilities using the *ggplot2* package [@Wickham2016-dg]. To facilitate this, *gratia* provides functions for representing the model components as objects using *tidy* principles, which are suitable for plotting with *ggplot2* or manipulation by packages in the *tidyverse* [e.g. @Wickham2023-uj]. This functionality allows for high-level plotting using the `draw()` method, as well as easily customisable plot creation using lower-level functionality.
 
 Taking a Bayesian approach to smoothing with penalized splines [@Kimeldorf1970-cn; @Wahba1983-mi; @Silverman1985-kw; @Wahba1985-bw; see @Miller2019-nf for a summary], it can be shown that GAMs fitted by *mgcv* are an empirical Bayesian model with an improper multivariate normal prior on the basis function coeficients. Samples from the posterior distribution of these models can be used to estimate the uncertainty in quantities derived from a GAM. This can be invaluable in applied research, where, for example, a quantity of interest may arise as an operation on predictions from the model. *gratia* provides functions for sampling from the posterior distribution of estimated smooths and from the model as a whole, where sampling can include the uncertainty in the estimated coefficients (`fitted_samples()`), the sampling uncertainty of the response (`predicted_samples()`), or both (`posterior_samples()`). By default, a Gaussian approximation to the posterior distribution is used, but a simple Metropolis Hasting sampler can be substituted (using `mgcv::gam.mh()`), which has better performance when the posterior is not well approximated by a Gaussian approximation.
 
[…]
