### 2024-11-21 b88cdc9
add comment regarding what the proj4 string is in the code as per #304
