### 2024-11-21 63f9f00
add a paragraph describing what the plots produced by draw show; #304
