+++ paper.Rmd
@@ -164,13 +164,15 @@ This model has much better model diagnostics although some large residuals remai
 appraise(m2)
 ```
 
-*gratia* can handle distributional GAMs fitted with *mgcv* and also *GJRM*'s `gamlss()`. Below, the estimated smooths from `m2` are plotted using `draw()`
+*gratia* can handle distributional GAMs fitted with *mgcv*, and also those fitted using *GJRM*'s `gamlss()`. Below, the estimated smooths from `m2` are plotted using `draw()`
 ```{r m2-draw, fig.show = "hide", dependson = "fit-chl-lsslmodel", cache = TRUE, fig.pos = "t!", out.width = ".8\\linewidth"}
 # Define the coordinate system to use for plotting on a map
 crs <- "+proj=ortho +lat_0=20 +lon_0=-40"
 draw(m2, crs = crs, default_crs = 4326, dist = 0.05, rug = FALSE)
 ```
-Here, we see a specialised plot drawn for spline-on-the-sphere smooths $f(\mathtt{lat}_i,\mathtt{lon}_i)$ (Figure \ref{fig:m2-draw}), which uses `ggplot2::coord_sf()` and functionality from the *sf* package [@Pebesma2018-ws; @Pebesma2023-fe] to visualise the smooth via an orthographic projection.
+The plots produced by `draw()` from a fitted model are known as *partial effect plots* (Figure \ref{fig:m2-draw}), which show the component contributions, on the link scale, of each model term to the linear predictor. The y axis on these plots is typically centred around 0 due to most smooths having a sum-to-zero identifiability constraint applied to them. This constraint is what allows the model to include multiple smooths and remain identifiable. These plots allow you to read off the contributions of each smooth to the fitted response (on the link scale); they show link-scale *predictions* of the response for each smooth, conditional upon all other terms in the model, including any parametric effects and the intercept, having zero contribution. In the paralance of the *marginaleffects* package, these plots show *adjusted predictions*, just where the adjustment includes setting the contribution of *all* other model terms to the predicted value to zero. For *partial derivatives* (what *marginaleffects* would call a *marginal effect* or slope), *gratia* provides the `derivatives()`.
+
+Here, we see a specialised plot drawn for spline-on-the-sphere smooths $f(\mathtt{lat}_i,\mathtt{lon}_i)$ (Figure \ref{fig:m2-draw}, left-hand column), which uses `ggplot2::coord_sf()` and functionality from the *sf* package [@Pebesma2018-ws; @Pebesma2023-fe] to visualise the smooth via an orthographic projection.
 
 ```{r m2-draw, echo = FALSE, fig.height = 12, fig.width = 15, fig.cap = "\\label{fig:m2-draw}Estimated smooth functions for the distributional GAM, `m2`, fitted to the ocean chlorophyll *a* data. The first row of plots is for the linear predictor of the conditional mean chlorophyll *a*, while the second and third rows are for the conditional power parameter and conditional scale, respectively. The shaded ribbons are 95% Bayesian credible intervals."}
 ```
