### 2024-11-21 990a5c8
add a state of the field section as per #304
