+++ paper.Rmd
@@ -105,6 +105,12 @@ Taking a Bayesian approach to smoothing with penalized splines [@Kimeldorf1970-c
 
 The teaching of GAMs can benefit from visualisation of the spline basis functions and associated penalty matrices. *gratia* provides this functionality via `basis()` and `penalty()`, which can be applied either to a smooth specification (e.g. `s(x, z, bs = "ds")`) or to a fitted GAM (see Figure \ref{fig:basis-funs}). These functions expose functionality already available in *mgcv*, but supply outputs in a tidy format, which makes access to these features more intuitive than the original implementations in *mgcv*. Additional utility functions are provided, for example: `model_constant()`, `edf()`, `model_edf()`, `overview()`, and `inv_link()`, which extract the model intercept term (or terms), the effective degrees of freedom of individual smooths and the overall model, shows a summary of the fitted GAM, and extracts the inverse of the link function(s) used, respectively.
 
+# State of the field
+
+Several R packages provide similar functionality to that of *gratia*. Notably, the *mgcViz* package [@Fasiolo2020-nz] provides sophisticated capabilities for plotting estimated smooths for models fitted by *mgcv* and *qgam* [@], and model diagnostics plots. A particular feature of *mgcViz* is that it was designed to be scalable, easily handling models fitted to data with millions of observations, a use case not currently well handled by *gratia*. *tidymv* [@tidymv-2023] and its succesor, *tidygam* [@tidygam-2023], provide plots of estimated smooths and model predictions respectively, as well as differences of smooths, but the focus is on univariate smooths, inconstrast to *gratia*'s ability to plot multivariate and specialist smooths (e.g. soap film smooths). Like *gratia*, the *itsadug* package [@itsadug-2022] is motivated to make working with estimated GAMs and related models easier. *itsadug* uses base graphics for plotting, but in place of partial effects plots, produces plots of adjusted predictions, and while useful for a broad range of models, is especially focused on GAMs fitted to longitudinal data with random effects.
+
+Areas where *gratia* stands out from these other packages are i) the consistent functionality for range of posterior simulations (*mgcViz* has some support for simulating new response values) as illustrated below, ii) support for computing derivatives of smooths, partial derivatives, and derivatives on the response scale (i.e. conditional derivatives), iii) tools for learning or teaching how GAMs work, iv) the range of utility functions that make working with GAMs easier, and v) the simple and consistent *tidy* interface to all functionality.
+
 # Example usage
 
 In this short example, I illustrate a few of the features of *gratia* using a data set of sea surface chlorophyll *a* measurements at a number of locations in the Atlantic Ocean, whose spatial locations are given as geographical coordinates (`lat` and `lon`), plus two additional covariates; `bathy`, the depth of the ocean, in metres, at the sampling location, and `jul.day`, the day of the year in which the observation was made. These data are in the `chl` dataset provided by the *gamair* package accompanying @Wood2017-qi.
@@ -170,7 +176,7 @@ appraise(m2)
 crs <- "+proj=ortho +lat_0=20 +lon_0=-40"
 draw(m2, crs = crs, default_crs = 4326, dist = 0.05, rug = FALSE)
 ```
[…]
