+++ paper.bib
@@ -353,3 +353,52 @@
   issn      = "0090-5364",
   language  = "en"
 }
+
+@article{Arel-BundockGreiferHeiss:2024,
+    title = {How to Interpret Statistical Models Using {marginaleffects} in {R} and {Python}},
+    author = {Vincent Arel-Bundock and Noah Greifer and Andrew Heiss},
+    year = {Forthcoming},
+    journal = {Journal of Statistical Software}}
+
+@Manual{itsadug-2022,
+    title = {{itsadug}: Interpreting Time Series and Autocorrelated
+      Data Using GAMMs},
+    author = {Jacolien {van Rij} and Martijn Wieling and R. Harald
+      Baayen and Hedderik {van Rijn}},
+    year = {2022},
+    note = {R package version 2.4.1},
+  }
+
+@Manual{tidygam-2023,
+  title = {tidygam: Tidy Prediction and Plotting of Generalised Additive Models},
+  author = {Stefano Coretta},
+  year = {2023},
+  note = {https://github.com/stefanocoretta/tidygam,
+https://stefanocoretta.github.io/tidygam/},
+}
+
+@Manual{tidymv-2023,
+  title = {tidymv: Tidy Model Visualisation for Generalised Additive Models},
+  author = {Stefano Coretta},
+  year = {2023},
+  note = {https://github.com/stefanocoretta/tidygam,
+https://stefanocoretta.github.io/tidygam/},
+}
+
+@ARTICLE{Fasiolo2020-nz,
+  title     = "Scalable Visualization Methods for Modern Generalized Additive
+               Models",
+  author    = "Fasiolo, Matteo and Nedellec, Raphaël and Goude, Yannig and Wood,
+               Simon N",
+  journal   = "Journal of computational and graphical statistics: a joint
+               publication of American Statistical Association, Institute of
+               Mathematical Statistics, Interface Foundation of North America",
+  publisher = "Taylor \& Francis",
+  volume    =  29,
+  number    =  1,
+  pages     = "78--86",
+  month     =  jan,
+  year      =  2020,
+  doi       = "10.1080/10618600.2019.1629942",
+  issn      = "1061-8600"
+}
