### 2024-11-21 15fa482
add a conclusion as per #304 and an acknowledgments section
