+++ paper.Rmd
@@ -36,7 +36,7 @@ A GAM has the form
 y_i &\sim    \mathcal{D}(\mu_i, \phi) \\
 g(\mu_i) &=  \mathbf{A}_i\boldsymbol{\gamma} + \sum_{j=1} f_j(x_{ji})
 \end{align*}
-where observations $y_i$ are assumed to be conditionally distributed $\mathcal{D}$ with expectation $\mathbb{E}(y_i) = \mu_i$ and dispersion $\phi$. The expectation of $y_i$ is given by a linear predictor of strictly parametric terms, whose model matrix is $\mathbf{A}_i$ with parameters $\boldsymbol{\gamma}$, plus a sum of smooth functions of $j = 1, \dots, J$ covariates $f_j()$. $g()$ is a link function mapping values on the linear predictor to the scale of the response.
+where observations $y_i$ are assumed to be conditionally distributed $\mathcal{D}$ with expectation $\mathbb{E}(y_i) = \mu_i$ and dispersion parameter $\phi$. The expectation of $y_i$ is given by a linear predictor of strictly parametric terms, whose model matrix is $\mathbf{A}_i$ with parameters $\boldsymbol{\gamma}$, plus a sum of smooth functions of $j = 1, \dots, J$ covariates $f_j()$. $g()$ is a link function mapping values on the linear predictor to the scale of the response.
 
 The smooth functions $f_j$ are represented in the GAM using penalised splines, which are themselves formed as weighted sums of basis functions, $b_k()$, [@De-Boor2001-vg] e.g.
 $$
@@ -44,9 +44,9 @@ f_j(x_{ij}) = \sum_{k=1}^{K} \beta_{jk} b_{jk}(x_{ij})
 $$
 for a univariate spline. The weights, $\beta_k$, are model coefficients to be estimated alongside $\boldsymbol{\gamma}$. To avoid overfitting, estimates $\hat{\beta}_{jk}$ and $\hat{\boldsymbol{\gamma}}$ are sought to minimise the penalised log-likelihood of the model
 $$
-\mathcal{L}(\boldsymbol{\beta}) = \ell(\boldsymbol{\beta}) - \frac{1}{2\phi} \sum_{j} \lambda_{j} \boldsymbol{\beta}^{\mathsf{T}}_j \mathbf{S}_j \boldsymbol{\beta}_j
+\ell_p(\boldsymbol{\beta}) = \ell(\boldsymbol{\beta}) - \frac{1}{2\phi} \sum_{j} \lambda_{j} \boldsymbol{\beta}^{\mathsf{T}}_j \mathbf{S}_j \boldsymbol{\beta}_j
 $$
-where $\ell$ is the log likelihood of the data at the parameter estimates, $\mathbf{S}_j$ are penalty matrices and $\lambda_{j}$ are smoothing parameters associated with each smooth. Note that $\boldsymbol{\beta}$ now contains the coefficients $\boldsymbol{\gamma}$ and $\beta_{jk}$. $\boldsymbol{\beta}^{\mathsf{T}}_j \mathbf{S}_j \boldsymbol{\beta}_j$ measures the wiggliness of $f_j$, which, with the default penalty, is the integrated squared second derivative of $f_j$. The smoothing parameters, $\boldsymbol{\lambda}$, control the trade-off between fit to the data and the complexity of the estimated functions.
[…]
