### 2024-11-21 a5fe939
more tweaks to rendering and modifications for #296 re phi or varphi
