### 2024-11-21 7634689
try to fix an issue with sizing of colourbars
