+++ R/smooth-estimates.R
@@ -1621,7 +1621,7 @@
       fill = guide_colourbar(
         title = guide_title,
         direction = "vertical",
-        barheight = grid::unit(3, "lines") #grid::unit(0.25, "npc")
+        barheight = grid::unit(5, "lines") #grid::unit(0.25, "npc")
       ),
       x = guide_axis(angle = angle)
     )
@@ -1766,7 +1766,7 @@
       fill = guide_colourbar(
         title = guide_title,
         direction = "vertical",
-        barheight = grid::unit(3, "lines") #grid::unit(0.25, "npc")
+        barheight = grid::unit(5, "lines") #grid::unit(0.25, "npc")
       ),
       x = guide_axis(angle = angle)
     )
@@ -1929,7 +1929,7 @@
       fill = guide_colourbar(
         title = guide_title,
         direction = "vertical",
-        barheight = grid::unit(3, "lines") #grid::unit(0.25, "npc")
+        barheight = grid::unit(5, "lines") #grid::unit(0.25, "npc")
       ),
       x = guide_axis(angle = angle)
     )
@@ -2551,7 +2551,7 @@
     guides(
       fill = guide_colourbar(
         title = guide_title, direction = "vertical",
-        barheight = grid::unit(3, "lines") #grid::unit(0.25, "npc")
+        barheight = grid::unit(5, "lines") #grid::unit(0.25, "npc")
       ),
       x = guide_axis(angle = angle)
     )
@@ -2695,7 +2695,7 @@
       fill = guide_colourbar(
         title = guide_title,
         direction = "vertical",
-        barheight = grid::unit(3, "lines") #grid::unit(0.25, "npc")
+        barheight = grid::unit(5, "lines") #grid::unit(0.25, "npc")
       ),
       x = guide_axis(angle = angle)
     )
