### 2024-11-21 b75bfc3
address @dill's comments from #296
