+++ paper.bib
@@ -111,7 +111,7 @@
 }
 
 @ARTICLE{Burkner2017-ms,
-  title    = "brms: An {R} Package for Bayesian Multilevel Models Using Stan",
+  title    = "{brms}: An {R} Package for Bayesian Multilevel Models Using Stan",
   author   = "Bürkner, Paul-Christian",
   journal  = "Journal of Statistical Software, Articles",
   volume   =  80,
@@ -124,7 +124,7 @@
 }
 
 @Manual{Marra2023-gjrm,
-    title = {GJRM: Generalised Joint Regression Modelling},
+    title = {{GJRM}: Generalised Joint Regression Modelling},
     pages = {R package version 0.2-6.4},
     year = {2023},
     author = {Giampiero Marra and Rosalba Radice},
@@ -212,7 +212,7 @@
 }
 
 @INCOLLECTION{Duchon1977-jr,
-  title     = "Splines minimizing rotation-invariant semi-norms in Sobolev
+  title     = "Splines minimizing rotation-invariant semi-norms in {S}obolev
                spaces",
   author    = "Duchon, Jean",
   booktitle = "Constructive Theory of Functions of Several Variables",
@@ -277,8 +277,7 @@
 }
 
 @ARTICLE{Wood2003-qy,
-  title     = "Thin plate regression splines: \textit{Thin Plate Regression
-               Splines}",
+  title     = "Thin plate regression splines",
   author    = "Wood, Simon N",
   journal   = "Journal of the Royal Statistical Society. Series B, Statistical
                methodology",
@@ -370,7 +369,7 @@
   }
 
 @Manual{tidygam-2023,
-  title = {tidygam: Tidy Prediction and Plotting of Generalised Additive Models},
+  title = {{tidygam}: Tidy Prediction and Plotting of Generalised Additive Models},
   author = {Stefano Coretta},
   year = {2023},
   note = {https://github.com/stefanocoretta/tidygam,
@@ -378,7 +377,7 @@ https://stefanocoretta.github.io/tidygam/},
 }
 
 @Manual{tidymv-2023,
-  title = {tidymv: Tidy Model Visualisation for Generalised Additive Models},
+  title = {{tidymv}: Tidy Model Visualisation for Generalised Additive Models},
   author = {Stefano Coretta},
   year = {2023},
   note = {https://github.com/stefanocoretta/tidygam,
@@ -390,9 +389,7 @@ https://stefanocoretta.github.io/tidygam/},
                Models",
   author    = "Fasiolo, Matteo and Nedellec, Raphaël and Goude, Yannig and Wood,
                Simon N",
-  journal   = "Journal of computational and graphical statistics: a joint
-               publication of American Statistical Association, Institute of
-               Mathematical Statistics, Interface Foundation of North America",
+  journal   = "{J}ournal of {C}omputational and {G}raphical {S}tatistics",
   publisher = "Taylor \& Francis",
   volume    =  29,
   number    =  1,
