### 2024-11-21 gavinsimpson/gratia#296 @gavinsimpson
@dill Thanks all the comments and suggestions and importantly taking the time to talk through some of the prior / null space things we both touch on in the comments above.

I've addressed all your comments (see above) bar 1:

> It would be nice to have a legend in Figure 1 so readers can relate the basis functions to the values in the penalty matrix (and e.g., see that the most wiggly function has the highest penalty)? Open to push-back on this, if you want to just give default plots from gratia

Done now. The basis function plot was not a plot that was generated by one of *gratia*'s plots, the penalty plot was but the basis function one needed some level of customisation anyway, so I added labelled lines.

I'll push this and then GH Actions will render the pdf to check. Will need to be looking in the joss-paper branch.
