### 2024-11-21 1627635
implement the last of @dill's suggestions; labelling the BFs in figure 1
