### 2024-11-21 9a01193
fix the labels on figure 1a
