### 2024-11-21 b590153
few tweaks to fix typos; add a missing reference
