+++ paper.bib
@@ -91,7 +91,7 @@
   author        = "Miller, David L",
   journal       = "arXiv [stat.ME]",
   month         =  feb,
-  year          =  2019,
+  year          =  2021,
   archivePrefix = "arXiv",
   primaryClass  = "stat.ME",
   eprint        = "1902.01330"
@@ -398,4 +398,16 @@ https://stefanocoretta.github.io/tidygam/},
   year      =  2020,
   doi       = "10.1080/10618600.2019.1629942",
   issn      = "1061-8600"
+}
+@ARTICLE{Fasiolo2020-pr,
+  title     = "Fast Calibrated Additive Quantile Regression",
+  author    = "Fasiolo, Matteo and Wood, Simon N and Zaffran, Margaux and
+               Nedellec, Raphaël and Goude, Yannig",
+  journal   = "Journal of the American Statistical Association",
+  publisher = "Taylor \& Francis",
+  pages     = "1--11",
+  month     =  feb,
+  year      =  2020,
+  doi       = "10.1080/01621459.2020.1725521",
+  issn      = "0162-1459"
 }
