### 2024-11-21 gavinsimpson/gratia#304 @gavinsimpson
@vankesteren I have now addressed all the points raised in your review comments above, with the exception of the one about loading packages as discussed in this thread.
