+++ paper.bib
@@ -128,6 +128,7 @@
     pages = {R package version 0.2-6.4},
     year = {2023},
     author = {Giampiero Marra and Rosalba Radice},
+    doi = "10.32614/CRAN.package.GJRM"
 }
 
 @ARTICLE{Pedersen2019-ff,
@@ -303,7 +304,8 @@
   number    =  2,
   pages     = "495--502",
   year      =  1970,
-  issn      = "0003-4851"
+  issn      = "0003-4851",
+  doi       = "10.1214/AOMS/1177697089"
 }
 
 @ARTICLE{Wahba1983-mi,
@@ -334,7 +336,8 @@
   number    =  1,
   pages     = "1--52",
   year      =  1985,
-  issn      = "1369-7412,0035-9246"
+  issn      = "1369-7412,0035-9246",
+  doi       = "10.1111/j.2517-6161.1985.tb01327.x"
 }
 
 @ARTICLE{Wahba1985-bw,
@@ -366,6 +369,7 @@
       Baayen and Hedderik {van Rijn}},
     year = {2022},
     note = {R package version 2.4.1},
+    doi = "10.32614/CRAN.package.itsadug"
   }
 
 @Manual{tidygam-2023,
@@ -374,6 +378,7 @@
   year = {2023},
   note = {https://github.com/stefanocoretta/tidygam,
 https://stefanocoretta.github.io/tidygam/},
+  doi = "10.32614/CRAN.package.tidygam"
 }
 
 @Manual{tidymv-2023,
@@ -382,6 +387,7 @@ https://stefanocoretta.github.io/tidygam/},
   year = {2023},
   note = {https://github.com/stefanocoretta/tidygam,
 https://stefanocoretta.github.io/tidygam/},
+  doi = "10.32614/CRAN.package.tidymv"
 }
 
 @ARTICLE{Fasiolo2020-nz,
