+++ paper.bib
@@ -20,6 +20,7 @@
   month     =  may,
   year      =  2017,
   isbn      =  9781498728379,
+  doi       = "10.1201/9781315370279",
   language  = "en"
 }
 
@@ -31,6 +32,7 @@
   month     =  jun,
   year      =  1990,
   isbn      =  9781351445962,
+  doi       = "10.1201/9780203753781-6",
   language  = "en"
 }
 
@@ -94,7 +96,8 @@
   year          =  2021,
   archivePrefix = "arXiv",
   primaryClass  = "stat.ME",
-  eprint        = "1902.01330"
+  eprint        = "1902.01330",
+  doi           = "10.48550/arXiv.1902.01330"
 }
 
 @BOOK{De-Boor2001-vg,
