### 2024-11-21 fd9dd24
fix some things on the contirbuting guide
