### 2024-11-21 28952b1
use the default barheight on multivariate smooths
