+++ tests/testthat/_snaps/backward/response-derivatives.md
@@ -26,15 +26,15 @@
       # A tibble: 50 x 9
           .row .focal .derivative .lower_ci .upper_ci    x0    x1      x2    x3
          <int> <chr>        <dbl>     <dbl>     <dbl> <dbl> <dbl>   <dbl> <dbl>
-       1     1 x2            217.     116.       316. 0.480 0.478 0.00127 0.537
+       1     1 x2            218.     116.       317. 0.480 0.478 0.00127 0.537
        2     2 x2            255.     128.       396. 0.480 0.478 0.0216  0.537
        3     3 x2            388.     282.       534. 0.480 0.478 0.0420  0.537
-       4     4 x2            478.     375.       618. 0.480 0.478 0.0624  0.537
+       4     4 x2            478.     375.       619. 0.480 0.478 0.0624  0.537
        5     5 x2            618.     485.       759. 0.480 0.478 0.0828  0.537
-       6     6 x2            668.     458.       872. 0.480 0.478 0.103   0.537
-       7     7 x2            617.     292.       929. 0.480 0.478 0.123   0.537
-       8     8 x2            388.     -23.4      785. 0.480 0.478 0.144   0.537
-       9     9 x2           -169.    -546.       197. 0.480 0.478 0.164   0.537
-      10    10 x2           -884.   -1312.      -499. 0.480 0.478 0.185   0.537
+       6     6 x2            668.     457.       872. 0.480 0.478 0.103   0.537
+       7     7 x2            617.     293.       928. 0.480 0.478 0.123   0.537
+       8     8 x2            387.     -22.0      786. 0.480 0.478 0.144   0.537
+       9     9 x2           -169.    -544.       196. 0.480 0.478 0.164   0.537
+      10    10 x2           -882.   -1311.      -497. 0.480 0.478 0.185   0.537
       # i 40 more rows
 
