+++ tests/testthat/_snaps/central/response-derivatives.md
@@ -31,10 +31,10 @@
        3     3 x2            379.     270.       528. 0.480 0.478 0.0420  0.537
        4     4 x2            505.     406.       642. 0.480 0.478 0.0624  0.537
        5     5 x2            616.     488.       753. 0.480 0.478 0.0828  0.537
-       6     6 x2            690.     479.       900. 0.480 0.478 0.103   0.537
-       7     7 x2            625.     274.       963. 0.480 0.478 0.123   0.537
-       8     8 x2            345.     -61.0      744. 0.480 0.478 0.144   0.537
+       6     6 x2            691.     479.       899. 0.480 0.478 0.103   0.537
+       7     7 x2            625.     275.       963. 0.480 0.478 0.123   0.537
+       8     8 x2            345.     -61.5      744. 0.480 0.478 0.144   0.537
        9     9 x2           -166.    -576.       227. 0.480 0.478 0.164   0.537
-      10    10 x2           -856.   -1295.      -464. 0.480 0.478 0.185   0.537
+      10    10 x2           -857.   -1295.      -464. 0.480 0.478 0.185   0.537
       # i 40 more rows
 
