+++ tests/testthat/_snaps/forward/response-derivatives.md
@@ -26,15 +26,15 @@
       # A tibble: 50 x 9
           .row .focal .derivative .lower_ci .upper_ci    x0    x1      x2    x3
          <int> <chr>        <dbl>     <dbl>     <dbl> <dbl> <dbl>   <dbl> <dbl>
-       1     1 x2            232.     145.       321. 0.480 0.478 0.00127 0.537
-       2     2 x2            293.     185.       420. 0.480 0.478 0.0216  0.537
+       1     1 x2            232.     145.       320. 0.480 0.478 0.00127 0.537
+       2     2 x2            292.     185.       420. 0.480 0.478 0.0216  0.537
        3     3 x2            366.     258.       511. 0.480 0.478 0.0420  0.537
-       4     4 x2            484.     381.       625. 0.480 0.478 0.0624  0.537
-       5     5 x2            589.     463.       726. 0.480 0.478 0.0828  0.537
-       6     6 x2            674.     454.       891. 0.480 0.478 0.103   0.537
-       7     7 x2            619.     298.       927. 0.480 0.478 0.123   0.537
-       8     8 x2            378.     -26.5      774. 0.480 0.478 0.144   0.537
-       9     9 x2           -129.    -547.       282. 0.480 0.478 0.164   0.537
+       4     4 x2            484.     381.       626. 0.480 0.478 0.0624  0.537
+       5     5 x2            588.     462.       725. 0.480 0.478 0.0828  0.537
+       6     6 x2            674.     455.       892. 0.480 0.478 0.103   0.537
+       7     7 x2            619.     299.       928. 0.480 0.478 0.123   0.537
+       8     8 x2            378.     -24.7      775. 0.480 0.478 0.144   0.537
+       9     9 x2           -129.    -546.       283. 0.480 0.478 0.164   0.537
       10    10 x2           -871.   -1307.      -481. 0.480 0.478 0.185   0.537
       # i 40 more rows
 
