### 2024-11-21 ea44b11
complete switch back to the default barheight
