### 2024-11-25 dfc6c8e
add conditional_values(), a simpler version of marginaleffects::plot_predictions(). Closes #300
