### 2024-11-25 ad96df3
add a subsetting function for conditional_values() as per #300
