+++ man/conditional_values.Rd
@@ -1,0 +1,123 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/conditional-values.R
+\name{conditional_values}
+\alias{conditional_values}
+\alias{conditional_values.gam}
+\title{Generated predicted values from a GAM, conditional upon supplied values of
+covariates}
+\usage{
+conditional_values(
+  model,
+  condition = NULL,
+  data = NULL,
+  scale = c("response", "link", "linear_predictor"),
+  ...
+)
+
+\method{conditional_values}{gam}(
+  model,
+  condition = NULL,
+  data = NULL,
+  scale = c("response", "link", "linear_predictor"),
+  n_vals = 100,
+  ci_level = 0.95,
+  ...
+)
+}
+\arguments{
+\item{model}{a fitted GAM object.}
+
+\item{condition}{either a character vector or a list supplying the names of
+covariates, and possibly their values, to condition up. The order of the
+values determines how these are plotted via the \code{\link[=draw.conditional_values]{draw.conditional_values()}}
+method; the first element is mapped to the \emph{x} channel, the second element
+to the \emph{colour} channel, the third to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} \strong{if no
+fourth element is present}, if present, the fourth element is mapped to
+the rows and the third element is mapped to the columns of
+\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.}
+
+\item{data}{data frame of values at which to predict. If supplied overrides
+values supplied through \code{condition}.}
+
+\item{scale}{character; which scale should predictions be returned on?}
+
+\item{...}{arguments passed to \code{\link[=fitted_values]{fitted_values()}}.}
+
+\item{n_vals}{numeric; number of values to generate for numeric variables
+named in \code{condition}.}
+
+\item{ci_level}{numeric; a number on interval (0,1) giving the coverage for
+credible intervals.}
+}
+\description{
+Generated predicted values from a GAM, conditional upon supplied values of
+covariates
+}
+\examples{
+load_mgcv()
+df <- data_sim("eg1", seed = 2)
+m1 <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = df, method = "REML")
+
+# predictions conditional on values evenly spaced over x2, all other
+# variables in model are held at representative values
+cv <- conditional_values(
+  m1,
+  condition = "x2"
+)
+# plot
+cv |> draw()
+
+# as above but condition on `x1` also. When plotted, `x1` is mapped to the
+# colour channel, noting that it has been summarised using fivenum()
+cv <- conditional_values(
+  m1,
+  condition = c("x2", "x1")
+)
+# plot
+cv |> draw()
+
+# can pass `condition` a list, allowing for greater flexibility
+# For example, here we condition on all four variables in the model,
+# summarising:
+#   * `x1` at its deciles,
+#   * `x0  at its quartiles
+#   * `x3` at its mean a d mean +/- sd
+cv <- conditional_values(
+  m1,
+  condition = list("x2", x1 = "decile", x0 = "quartile", "threenum")
+)
+# plot
+cv |> draw()
+
+# some model terms can be exclude from the conditional predictions using the
+# `exclude` mechanism of `predict.gam`. Here we exclude the effects of
+# `s(x0)` and `s(x3)` from the conditional predictions. This, in effect,
+# treats these smooths as having **0** effect on the conditional predictions
+# of the response, even though the two smooths conditioned on (`s(x2)` and
[…]
