+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method("[",conditional_values)
 S3method("[",eval_sm)
 S3method("[",parametric_effects)
 S3method("[",smooth_estimates)
@@ -31,6 +32,7 @@ S3method(basis_size,mgcv.smooth)
 S3method(boundary,gam)
 S3method(boundary,soap.film)
 S3method(coef,scam)
+S3method(conditional_values,gam)
 S3method(confint,fderiv)
 S3method(confint,gam)
 S3method(confint,gamm)
@@ -62,6 +64,7 @@ S3method(dispersion,gam)
 S3method(dispersion,glm)
 S3method(draw,basis)
 S3method(draw,compare_smooths)
+S3method(draw,conditional_values)
 S3method(draw,derivatives)
 S3method(draw,difference_smooth)
 S3method(draw,evaluated_parametric_term)
@@ -317,6 +320,7 @@ export(by_variable)
 export(check_is_mgcv_smooth)
 export(compare_smooths)
 export(concrvity)
+export(conditional_values)
 export(data_combos)
 export(data_sim)
 export(data_slice)
@@ -482,6 +486,7 @@ importFrom(ggplot2,label_value)
 importFrom(ggplot2,labs)
 importFrom(ggplot2,layer_scales)
 importFrom(ggplot2,lims)
+importFrom(ggplot2,position_dodge)
 importFrom(ggplot2,scale_colour_continuous)
 importFrom(ggplot2,scale_colour_discrete)
 importFrom(ggplot2,scale_colour_hue)
@@ -542,6 +547,7 @@ importFrom(rlang,inject)
 importFrom(rlang,parse_expr)
 importFrom(rlang,set_names)
 importFrom(stats,IQR)
+importFrom(stats,as.formula)
 importFrom(stats,binomial)
 importFrom(stats,coef)
 importFrom(stats,confint)
@@ -552,6 +558,7 @@ importFrom(stats,dnorm)
 importFrom(stats,dpois)
 importFrom(stats,family)
 importFrom(stats,fitted)
+importFrom(stats,fivenum)
 importFrom(stats,formula)
 importFrom(stats,median)
 importFrom(stats,model.frame)
@@ -573,6 +580,7 @@ importFrom(stats,rnbinom)
 importFrom(stats,rnorm)
 importFrom(stats,rpois)
 importFrom(stats,runif)
+importFrom(stats,sd)
 importFrom(stats,setNames)
 importFrom(stats,simulate)
 importFrom(stats,symnum)
@@ -606,6 +614,7 @@ importFrom(tidyselect,eval_select)
 importFrom(tidyselect,everything)
 importFrom(tidyselect,last_col)
 importFrom(tidyselect,matches)
+importFrom(tidyselect,where)
 importFrom(tools,toTitleCase)
 importFrom(utils,capture.output)
 importFrom(utils,combn)
