### 2024-11-25 a10e9bb
fit cnorm model with gam not bam as I'm seeing a step failure warning on macos x
