### 2024-11-25 a3c71cd
add some basic tests for conditional_values
