### 2024-11-25 01753cf
document new conditional_values()
