### 2024-11-25 808bd5c
fix a typo in the example
