+++ R/conditional-values.R
@@ -52,7 +52,7 @@
 #' #   * `x3` at its mean a d mean +/- sd
 #' cv <- conditional_values(
 #'   m1,
-#'   condition = list("x2", x1 = "decile", x0 = "quartile", "threenum")
+#'   condition = list("x2", x1 = "decile", x0 = "quartile", x3 = "threenum")
 #' )
 #' # plot
 #' cv |> draw()
