### 2024-11-25 705cf65
document() and update example reference output
