### 2024-11-25 fc6c951
tweak an example
