### 2024-11-25 3556614
expand documentation for conditional_values()
