+++ NEWS.md
@@ -1,4 +1,4 @@
-# gratia
+# gratia (development version)
 
 ## New features
 
