### 2024-11-25 c06775b
tweak the getting started vignette
