+++ vignettes/getting-started.bib
@@ -1,0 +1,5 @@
+@article{Arel-BundockGreiferHeiss:2024,
+    title = {How to Interpret Statistical Models Using {marginaleffects} in {R} and {Python}},
+    author = {Vincent Arel-Bundock and Noah Greifer and Andrew Heiss},
+    year = {Forthcoming},
+    journal = {Journal of Statistical Software}}
