### 2024-11-25 8e7de96
add a new example of posterior simulation from a distributional GAM
