### 2024-11-25 6f8dabd
fix a couple of typos
