### 2024-12-13 de73124
fix issue with the fix for #323
