### 2024-12-13 b01bc6c
moar tests for conditional_values()
