+++ tests/testthat/_snaps/backward/response-derivatives.md
@@ -32,9 +32,9 @@
        4     4 x2            478.     375.       619. 0.480 0.478 0.0624  0.537
        5     5 x2            618.     485.       759. 0.480 0.478 0.0828  0.537
        6     6 x2            668.     457.       872. 0.480 0.478 0.103   0.537
-       7     7 x2            617.     293.       928. 0.480 0.478 0.123   0.537
+       7     7 x2            617.     292.       928. 0.480 0.478 0.123   0.537
        8     8 x2            387.     -22.0      786. 0.480 0.478 0.144   0.537
-       9     9 x2           -169.    -544.       196. 0.480 0.478 0.164   0.537
-      10    10 x2           -882.   -1311.      -497. 0.480 0.478 0.185   0.537
+       9     9 x2           -169.    -545.       196. 0.480 0.478 0.164   0.537
+      10    10 x2           -882.   -1311.      -496. 0.480 0.478 0.185   0.537
       # i 40 more rows
 
