+++ tests/testthat/_snaps/confint-methods.md
@@ -20,15 +20,15 @@
     # A tibble: 25 x 4
        term   lower      est  upper
        <chr>  <dbl>    <dbl>  <dbl>
-     1 s(x2)  -7.99   0.0645   8.12
-     2 s(x2) -25.0  -15.9     -6.75
+     1 s(x2)  -7.96   0.0645   8.09
+     2 s(x2) -24.9  -15.9     -6.78
      3 s(x2) -37.8  -28.6    -19.3 
      4 s(x2) -45.3  -36.8    -28.4 
-     5 s(x2) -48.0  -40.3    -32.7 
-     6 s(x2) -47.4  -39.5    -31.5 
-     7 s(x2) -44.8  -35.7    -26.6 
-     8 s(x2) -39.9  -30.4    -20.9 
+     5 s(x2) -47.9  -40.3    -32.7 
+     6 s(x2) -47.4  -39.5    -31.6 
+     7 s(x2) -44.7  -35.7    -26.7 
+     8 s(x2) -39.9  -30.4    -21.0 
      9 s(x2) -33.5  -24.6    -15.8 
-    10 s(x2) -27.0  -19.1    -11.2 
+    10 s(x2) -26.9  -19.1    -11.2 
     # i 15 more rows
 
