+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: gratia
-Version: 0.9.2.9013
+Version: 0.9.2.9014
 Title: Graceful 'ggplot'-Based Graphics and Other Functions for GAMs Fitted Using 'mgcv'
 Authors@R: c(person(given = "Gavin L.", family = "Simpson",
                     email = "ucfagls@gmail.com",
@@ -33,7 +33,8 @@ Imports:
     cli,
     nlme,
     ggokabeito,
-    withr
+    withr,
+    scales
 Suggests:
     gamm4,
     lme4,
