+++ NAMESPACE
@@ -418,7 +418,9 @@ export(variance_comp)
 export(vars_from_label)
 export(which_smooths)
 export(worm_plot)
+importFrom(cli,format_error)
 importFrom(cli,pluralize)
+importFrom(cli,qty)
 importFrom(cli,style_dim)
 importFrom(cli,symbol)
 importFrom(dplyr,"%>%")
@@ -533,6 +535,7 @@ importFrom(rlang,"!!")
 importFrom(rlang,"%||%")
 importFrom(rlang,":=")
 importFrom(rlang,.data)
+importFrom(rlang,abort)
 importFrom(rlang,dots_list)
 importFrom(rlang,enexpr)
 importFrom(rlang,enquo)
@@ -546,6 +549,7 @@ importFrom(rlang,has_name)
 importFrom(rlang,inject)
 importFrom(rlang,parse_expr)
 importFrom(rlang,set_names)
+importFrom(scales,ordinal_format)
 importFrom(stats,IQR)
 importFrom(stats,as.formula)
 importFrom(stats,binomial)
