### 2024-12-19 2381bf1
suppress GJRM startup messages
