+++ man/draw.gamlss.Rd
@@ -49,7 +49,7 @@ Plots of smooths are not labelled with the linear predictor to which
 they belong.
 }
 \examples{
-if (require("GJRM", quietly = TRUE)) {
+if (suppressPackageStartupMessages(require("GJRM", quietly = TRUE))) {
   # follow example from ?GJRM::gamlss
   load_mgcv()
   suppressPackageStartupMessages(library("GJRM"))
