### 2024-12-19 #6962 @gavinsimpson
Done:

    Version 0.10.0 has been released and is on CRAN: https://github.com/gavinsimpson/gratia/releases/tag/v0.10.0
    https://zenodo.org/records/14531826 with DOI [10.5281/zenodo.14531826](https://doi.org/10.5281/zenodo.14531826)
