### 2024-12-20 #6962 @gavinsimpson
@lrnv Then I'm not sure what I was supposed to archive? The Zenodo archive is of the R package, which has a (neccessarily) different title to the paper.

So, I'm thinking I misunderstood what I was archiving? I thought it was the release &mdash; but that's not the branch with the paper.
