### 2024-12-20 #6962 @lrnv
You should archive the release as you did, but the title of the archive should match the title of the paper. You can edit a-posteriori the archive's title on zenodo.
