### 2024-12-20 #6962 @gavinsimpson
OK, that I can do; but I am not relicensing the R package as CC-BY isn't suitable for code. Isn't the CC BY license for the paper?
