### 2024-12-20 #6962 @lrnv
Nope, the licence written in the paper is the licence of the code -- thus MIT here
