### 2024-12-20 #6962 @lrnv
Btw, JOSS published the code, and even if it is a bit late to point it out, the paper you wrote is a bit long regarding our standards... 2300ish words instead of ~1000. No obligation since it's late, but if you feel like some paragraphs are too much please cut them out :)
