### 2024-12-20 #6962 @gavinsimpson
Where in the paper do I say the package is licensed CC BY?
