### 2024-12-20 #6962 @lrnv
> Where in the paper do I say the package is licensed CC BY?

You re right, this is automatic and should be fixed since we setted the archive. Let us see what the next generated pdf say
