### 2024-12-20 #6962 @lrnv
@gavinsimpson Yes that looks like it is the issue. Do you really ned both the `LICENSE` and `LICENSE.md` files ? Should'nt they contain the same thing ? Or at least the `LICENSE` file should contain the actual licence, right ?
