### 2024-12-20 #6962 @gavinsimpson
> Hum... Did not work. Maybe it is because you have both a LICENSE and LICENSE.md file ? Not sure

> Do you really need both the LICENSE and LICENSE.md files ? Should'nt they contain the same thing ? Or at least the LICENSE file should contain the actual licence, right ?

No, unfortunately.

The `LICENSE` is needed to meet CRAN's requirements when using the MIT (the `LICENSE` contains the copyright holder info that CRAN uses to supplement a generic MIT license).

GitHub doesn't understand any of this so best practice is to include the actual license via `LICENSE.md`

The `LICENSE` is the only file that gets added to the R package when being built from the sources. `LICENSE.md` is for everyone else to work around CRAN's finicky requirements.
