### 2024-12-20 #6962 @lrnv
Let me have a chat with other editors more accustomed to R and CRAN shananigans and come back to you. I'm starting to think that the current state of things is good to go, but let me check
