### 2024-12-20 #6962 @lrnv
Yep, by looking at other cases we agree that it's Ok. Alright, sorry for the fuzz
