### 2024-12-20 #6962 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
