### 2024-12-20 abe3bc6
fix a couple of typos
