### 2024-12-20 #6962 @gavinsimpson
@lrnv I have fixed the two typos I found after a final read through of the proofs and am now finished editing.

Thank you for handling my submission to JOSS
