### 2024-12-20 #6962 @crvernon
:wave: @gavinsimpson - I just need you to address the following before I move to accept this for publication:

- [ ] Please remove the "forthcoming" citation from the paper and bibliography.
- [ ] LINE 220:  The "b" in "bayesian" should be capitalized.   You can achieve this by placing curly brackets around characters you wish to maintain formatting of in your bib file.  Please also check all other references for correct capitalization as well.

After you have made these changes, let me know.  Thanks!
