### 2024-12-21 e4bc38c
fix two references
