+++ paper.bib
@@ -114,7 +114,7 @@
 }
 
 @ARTICLE{Burkner2017-ms,
-  title    = "{brms}: An {R} Package for Bayesian Multilevel Models Using Stan",
+  title    = "{brms}: An {R} Package for {B}ayesian Multilevel Models Using {Stan}",
   author   = "Bürkner, Paul-Christian",
   journal  = "Journal of Statistical Software, Articles",
   volume   =  80,
@@ -359,11 +359,21 @@
   language  = "en"
 }
 
-@article{Arel-BundockGreiferHeiss:2024,
-    title = {How to Interpret Statistical Models Using {marginaleffects} in {R} and {Python}},
-    author = {Vincent Arel-Bundock and Noah Greifer and Andrew Heiss},
-    year = {Forthcoming},
-    journal = {Journal of Statistical Software}}
+@ARTICLE{Arel-BundockGreiferHeiss:2024,
+  title     = "How to Interpret Statistical Models Using {marginaleffects} for
+               {R} and {P}ython",
+  author    = "Arel-Bundock, Vincent and Greifer, Noah and Heiss, Andrew",
+  journal   = "Journal of statistical software",
+  publisher = "Foundation for Open Access Statistic",
+  volume    =  111,
+  number    =  9,
+  pages     = "1--32",
+  month     =  nov,
+  year      =  2024,
+  doi       = "10.18637/jss.v111.i09",
+  issn      = "1548-7660,1548-7660",
+  language  = "en"
+}
 
 @Manual{itsadug-2022,
     title = {{itsadug}: Interpreting Time Series and Autocorrelated
