+++ paper.Rmd
@@ -26,7 +26,9 @@ journal: JOSS
 # Summary
 
 ```{r, setup, echo = FALSE, cache = FALSE, results = "hide"}
-pkgs <- c("mgcv", "gratia", "ggplot2", "geomtextpath", "dplyr", "tibble", "patchwork")
+pkgs <- c(
+  "mgcv", "gratia", "ggplot2", "geomtextpath", "dplyr", "tibble", "patchwork"
+)
 res <- vapply(pkgs, library, logical(1), character.only = TRUE,
   logical.return = TRUE)
 ```
@@ -131,9 +133,9 @@ loaded <- vapply(pkgs, library, logical(1L), logical.return = TRUE,
 data(chl, package = "gamair")
 ```
 
-A simple GAM for these data is to model the response (`chl`) with a spatial smooth of latitude (`lat`) and longitude (`lon`) as covariates. Here, I use a spline on the sphere (SOS) smoother built using a Duchon spline with second order derivative penalty [@Duchon1977-jr]. Additional terms included in the linear predictor are a smooth of the day of year of sample collection (`jul.day`) and a smooth of ocean depth (`bath`). The response is assumed to be conditionally Tweedie distributed, with the power parameter ($p$) of the distribution estimated during fitting. Model coefficients and smoothing parameters are estimated using restricted maximum likelihood [@Wood2011-kn]. To use parallel processing in some parts of the model fitting algorithm, the `nthreads` control parameter is set to 10 (set this lower if your machine doesn't have this many physical CPU cores).
+A simple GAM for these data is to model the response (`chl`) with a spatial smooth of latitude (`lat`) and longitude (`lon`) as covariates. Here, I use a spline on the sphere (SOS) smoother built using a Duchon spline with second order derivative penalty [@Duchon1977-jr]. Additional terms included in the linear predictor are a smooth of the day of year of sample collection (`jul.day`) and a smooth of ocean depth (`bath`). The response is assumed to be conditionally Tweedie distributed, with the power parameter ($p$) of the distribution estimated during fitting. Model coefficients and smoothing parameters are estimated using restricted maximum likelihood [@Wood2011-kn]. To use parallel processing in some parts of the model fitting algorithm, the `nthreads` control parameter is set to 8 (set this lower if your machine doesn't have this many physical CPU cores).
 ```{r fit-chl-model, cache = TRUE}
-ctrl <- gam.control(nthreads = 10)
+ctrl <- gam.control(nthreads = 8)
 m1 <- gam(
   chl ~ s(lat, lon, bs = "sos", m = -1, k = 150) +
     s(jul.day, bs = "cr", k = 20) +
