### 2024-12-21 #6962 @gavinsimpson
@crvernon

* [x] Please remove the "forthcoming" citation from the paper and bibliography.
    *This paper was just published so I have updated the reference in the paper with the new metadata.*
* [x] LINE 220: The "b" in "bayesian" should be capitalized. You can achieve this by placing curly brackets around characters you wish to maintain formatting of in your bib file. Please also check all other references for correct capitalization as well.
    *Fixed this plus another. All other references have been checked.*

Thanks
